refactor(template-compiler): merge TemplateBinder, AttrSyntaxTransformer and TemplateElementFactory into TemplateCompiler
